refactor dpm construction action routes
